requirements.txt
crewai==0.51.1
crewai-tools==0.8.3
langchain==0.2.13
langchain_core==0.2.30
langchain-community==0.2.12
langchain_google_genai==1.0.8
langchain-huggingface==0.0.3
langchain-groq==0.1.9
duckduckgo-search==6.2.6
python-dotenv==1.0.1
streamlit==1.34.0
transformers==4.44.0
numexpr==2.10.1
beautifulsoup4==4.12.3